Add DocumentSortByMapper to handle orderBy mapping for WebDavDocument queries
